+++ src/test/circular-types/fs-tree/file.ts
@@ -2,8 +2,7 @@ import {CaseStyle} from "../../../lib/_helpers/rename";
 import {DocumentType} from "../../../lib/document";
 import {IntegerType} from "../../../lib/integer";
 import {LiteralType} from "../../../lib/literal";
-import {Ucs2StringType} from "../../../lib/ucs2-string";
-import {FsNodeBase} from "./fs-node-base";
+import {$FsNodeBase, FsNodeBase} from "./fs-node-base";
 import {$FsNodeType, FsNodeType} from "./fs-node-type";
 
 export interface File extends FsNodeBase {
@@ -13,8 +12,8 @@ export interface File extends FsNodeBase {
 
 export const $File: DocumentType<File> = new DocumentType<File>({
   properties: {
+    ...$FsNodeBase.properties,
     type: {type: new LiteralType({type: $FsNodeType, value: FsNodeType.File})},
-    name: {type: new Ucs2StringType({maxLength: Infinity})},
     size: {type: new IntegerType()},
   },
   rename: CaseStyle.SnakeCase,