+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next
 
-- **[Feature]** Support lazy `Union` definition.
+- **[Feature]** Support lazy options definition. Instead of an option object, you can now pass a function returning an
+  option object. It will be evaluated only once one of its values is needed. This enables circular dependencies.
 - **[Internal]** Update build tools.
 
 # 0.5.0-alpha.10 (2017-08-11)